% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/im_cols.R
\name{im_cols}
\alias{im_cols}
\title{Pick color from imcolors palette}
\usage{
im_cols(...)
}
\arguments{
\item{...}{color name as string}
}
\value{
string
}
\description{
Pick color from imcolors palette
}
\examples{

}
