% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/im_cols_gg.R
\name{scale_color_im}
\alias{scale_color_im}
\title{Scale_color function for ggplot2}
\usage{
scale_color_im(palette = "main", discrete = TRUE, reverse = FALSE, ...)
}
\arguments{
\item{palette}{palette name from impalettes}

\item{discrete}{is discrete?}

\item{reverse}{reverse order}

\item{...}{}
}
\value{

}
\description{
Scale_color function for ggplot2
}
